requirements.txt
python-telegram-bot==13.13
openai==0.27.9
configparser=6.0.0
PyYAML=6.0.1